modify seele.NewMiner and related codes

+++ go-seele-one-shard/core/tx_pool.go
@@ -54,11 +54,11 @@ type TransactionPool struct {
 	pendingQueue  *pendingQueue
 	processingTxs map[common.Hash]struct{}
 	log           *log.SeeleLog
-	chainNum      uint
+	chainNum      uint64
 }
 
 // NewTransactionPool creates and returns a transaction pool.
-func NewTransactionPool(config TransactionPoolConfig, chain blockchain, chainNum uint) *TransactionPool {
+func NewTransactionPool(config TransactionPoolConfig, chain blockchain, chainNum uint64) *TransactionPool {
 	pool := &TransactionPool{
 		config:        config,
 		chain:         chain,
